+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Fixes
 
 - Don't use main thread for app hang screenshot and view hierarchy (#3547)
+- Finish TTID span when transaction finishes (#3610)
+
 
 ## 8.20.0